+++ src/lib/components/NostrElements/kindEvents/NoteActionButtuns/CustomReaction.svelte
@@ -1,11 +1,12 @@
 <script lang="ts">
   import Popover from "$lib/components/Elements/Popover.svelte";
   import { emojis } from "$lib/stores/stores";
-  import { SmilePlus } from "lucide-svelte";
+  import { RefreshCw, SmilePlus } from "lucide-svelte";
   import * as Nostr from "nostr-typedef";
   import split from "graphemesplit";
   import { clientTag } from "$lib/func/constants";
   import { lumiSetting } from "$lib/stores/globalRunes.svelte";
+  import EmojiListUpdate from "$lib/components/SettingsElements/EmojiListUpdate.svelte";
 
   interface Props {
     note: Nostr.Event | undefined;
@@ -137,33 +138,37 @@
         </button>
       </div>
 
-      {#if $emojis && $emojis.list.length > 0}
-        <div
-          class=" mt-1 border border-magnum-600 flex flex-wrap max-w-80 max-h-80 overflow-y-auto"
+      <div
+        class=" mt-1 border border-magnum-600 flex flex-wrap max-w-80 max-h-80 overflow-y-auto"
+      >
+        {#each $emojis.list as e, index}
+          {#if customReaction === "" || e[0]
+              .toLowerCase()
+              .includes(customReaction.replace(":", "").toLowerCase())}
+            <button
+              aria-label={`Select emoji ${e[0]}`}
+              onclick={() => handleClickEmoji(e)}
+              class="rounded-md border m-0.5 p-1 border-magnum-600 font-medium text-magnum-100 hover:opacity-75 active:opacity-50 text-sm"
+            >
+              {#if lumiSetting.get().showImg}
+                <img
+                  height="24px"
+                  loading="lazy"
+                  class="h-6 min-w-6 object-contain justify-self-center"
+                  src={e[1]}
+                  alt={e[0]}
+                  title={e[0]}
+                />{:else}{e[0]}{/if}
+            </button>
+          {/if}
+        {/each}
+        <EmojiListUpdate
+          buttonClass="ml-auto p-1 m-1 rounded-full  bg-magnum-500 hover:opacity-75 active:opacity-50"
         >
-          {#each $emojis.list as e, index}
-            {#if customReaction === "" || e[0]
-                .toLowerCase()
-                .includes(customReaction.replace(":", "").toLowerCase())}
-              <button
-                aria-label={`Select emoji ${e[0]}`}
-                onclick={() => handleClickEmoji(e)}
-                class="rounded-md border m-0.5 p-1 border-magnum-600 font-medium text-magnum-100 hover:opacity-75 active:opacity-50 text-sm"
-              >
-                {#if lumiSetting.get().showImg}
-                  <img
-                    height="24px"
-                    loading="lazy"
-                    class="h-6 min-w-6 object-contain justify-self-center"
-                    src={e[1]}
-                    alt={e[0]}
-                    title={e[0]}
-                  />{:else}{e[0]}{/if}
-              </button>
-            {/if}
-          {/each}
-        </div>
-      {/if}
+          <RefreshCw />
+        </EmojiListUpdate>
+      </div>
+
       {#if customReactionError}
         <div class="text-red-500 text-sm mt-1">
           {customReactionErrorMessage}
