+++ src/lib/components/SettingsElements/UpdateEmojiList.svelte
@@ -1,114 +1,29 @@
 <script lang="ts">
-  import * as Nostr from "nostr-typedef";
-  import {
-    createEmojiListFrom10030,
-    getDoukiList,
-    getQueryRelays,
-  } from "$lib/func/settings";
   import { formatAbsoluteDate } from "$lib/func/util";
-  import {
-    emojis,
-    loginUser,
-    nowProgress,
-    toastSettings,
-  } from "$lib/stores/stores";
+  import { emojis, loginUser, nowProgress } from "$lib/stores/stores";
   import Dialog from "../Elements/Dialog.svelte";
   import { _ } from "svelte-i18n";
 
-  import { createRxNostr } from "rx-nostr/src";
-  import { get, writable, type Writable } from "svelte/store";
-  import { verifier as cryptoVerifier } from "rx-nostr-crypto";
+  import { writable, type Writable } from "svelte/store";
   import { nip19 } from "nostr-tools";
 
-  import { lumiSetting, verifier } from "$lib/stores/globalRunes.svelte";
+  import { lumiSetting } from "$lib/stores/globalRunes.svelte";
   import { SmilePlus } from "lucide-svelte";
+  import EmojiListUpdate from "./EmojiListUpdate.svelte";
   interface Props {
     pubkey: string;
   }
 
-  let { pubkey = $bindable() }: Props = $props();
-  // export let emojiList: LumiEmoji | undefined;
   // svelte-ignore non_reactive_update
   let dialogOpen: Writable<boolean> = writable(false);
-  async function handleClickEmoji() {
-    const beforeEvent = $emojis?.event;
-
-    try {
-      const gotPubkey = await (
-        window.nostr as Nostr.Nip07.Nostr
-      ).getPublicKey();
-      if (gotPubkey) {
-        pubkey = gotPubkey;
-      }
-    } catch (error) {
-      console.log(error);
-    }
-    if (pubkey === "") {
-      $toastSettings = {
-        title: "Error",
-        description: "pubkey not found ",
-        color: "bg-red-500",
-      };
-      return;
-    }
-    const relays = await getQueryRelays(pubkey);
-    console.log(relays);
-    if (!relays) {
-      $toastSettings = {
-        title: "Error",
-        description: "relay list not found",
-        color: "bg-red-500",
-      };
-      return;
-    }
-    $nowProgress = true;
-    const filters: Nostr.Filter[] = [
-      { limit: 1, kinds: [10030], authors: [pubkey] },
-    ];
-    const pk = await getDoukiList(filters, relays);
-    // console.log(event);
-    if (!pk && !beforeEvent) {
-      $toastSettings = {
-        title: "Warning",
-        description: "emoji list not found",
-        color: "bg-red-500",
-      };
-      $nowProgress = false;
-      return;
-    }
-    const kind10030event =
-      !beforeEvent ||
-      beforeEvent.pubkey !== pk.event.pubkey ||
-      pk.event.created_at >= beforeEvent.created_at
-        ? pk.event
-        : beforeEvent;
-
-    const rxNostr = createRxNostr({
-      verifier: verifier.get() ?? cryptoVerifier,
-    });
-    const list = await createEmojiListFrom10030(
-      kind10030event,
-      rxNostr,
-      relays
-    );
-
-    $emojis = {
-      list: list,
-      updated: Math.floor(Date.now() / 1000),
-      event: kind10030event,
-    };
-
-    localStorage.setItem("lumiEmoji", JSON.stringify($emojis));
-
-    $nowProgress = false;
-  }
 </script>
 
-<button
-  disabled={$nowProgress}
-  class="h-10 ml-2 rounded-md bg-magnum-600 px-3 py-1 font-medium text-magnum-100 hover:opacity-75 active:opacity-50 disabled:opacity-25"
-  onclick={handleClickEmoji}>Emoji</button
-><time class="ml-2"
+<EmojiListUpdate
+  buttonClass="h-10 ml-2 rounded-md bg-magnum-600 px-3 py-1 font-medium text-magnum-100 hover:opacity-75 active:opacity-50 disabled:opacity-25"
+>
+  Emoji
+</EmojiListUpdate>
+<time class="ml-2"
   >{$_("settings.lastUpdated")}: {$emojis
     ? formatAbsoluteDate($emojis?.updated)
     : ""}</time
